requirements.txt
MouseInfo==0.1.3
packaging==24.1
pillow==11.0.0
pure-python-adb==0.3.0.dev0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
pytweening==1.2.0
setuptools==75.2.0
uiautomator==1.0.2
urllib3==2.2.3
